Crate pallet_atomic_swap

Atomic Swap

A pallet for atomically sending funds.

Config

Call

Pallet

Overview

A pallet for atomically sending funds from an origin to a target. A proof is used to allow the target to approve (claim) the swap. If the swap is not claimed within a specified duration of time, the sender may cancel it.

Interface

Dispatchable Functions

create_swap - called by a sender to register a new atomic swap

claim_swap - called by the target to approve a swap

cancel_swap - may be called by a sender after a specified duration

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

BalanceSwapAction

A swap action that only allows transferring balances.

PendingSwap

Pending atomic swap operation.

Traits

SwapAction

Definition of a pending atomic swap action. It contains the following three phrases:

Type Aliases

HashedProof

Hashed proof type.