pallet_atomic_swap::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

cancel_swap

Cancel an atomic swap. Only possible after the originally set duration has passed.

claim_swap

Claim an atomic swap.

create_swap

Register a new atomic swap, declaring an intention to send funds from origin to target on the current blockchain. The target can claim the fund using the revealed proof. If the fund is not claimed after duration blocks, then the sender can cancel the swap.