Type Alias pallet_atomic_swap::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn create_swap( origin: OriginFor<T>, target: T::AccountId, hashed_proof: HashedProof, action: T::SwapAction, duration: BlockNumberFor<T> ) -> DispatchResult

Register a new atomic swap, declaring an intention to send funds from origin to target on the current blockchain. The target can claim the fund using the revealed proof. If the fund is not claimed after duration blocks, then the sender can cancel the swap.

The dispatch origin for this call must be Signed.

target: Receiver of the atomic swap.

hashed_proof: The blake2_256 hash of the secret proof.

balance: Funds to be sent from origin.

duration: Locked duration of the atomic swap. For safety reasons, it is recommended that the revealer uses a shorter duration than the counterparty, to prevent the situation where the revealer reveals the proof too late around the end block.

pub fn claim_swap( origin: OriginFor<T>, proof: Vec<u8>, action: T::SwapAction ) -> DispatchResult

Claim an atomic swap.

The dispatch origin for this call must be Signed.

proof: Revealed proof of the claim.

action: Action defined in the swap, it must match the entry in blockchain. Otherwise the operation fails. This is used for weight calculation.

pub fn cancel_swap( origin: OriginFor<T>, target: T::AccountId, hashed_proof: HashedProof ) -> DispatchResult

Cancel an atomic swap. Only possible after the originally set duration has passed.

The dispatch origin for this call must be Signed.

target: Target of the original atomic swap.

hashed_proof: Hashed proof of the original atomic swap.