Crate pallet_atomic_swap

Atomic Swap

A module for atomically sending funds.

atomic_swap::Trait

Call

Module

Overview

A module for atomically sending funds from an origin to a target. A proof is used to allow the target to approve (claim) the swap. If the swap is not claimed within a specified duration of time, the sender may cancel it.

Interface

Dispatchable Functions

create_swap - called by a sender to register a new atomic swap

claim_swap - called by the target to approve a swap

cancel_swap - may be called by a sender after a specified duration

Structs

BalanceSwapAction
A swap action that only allows transferring balances.

Module
Module definition of atomic swap pallet.

PendingSwap
Pending atomic swap operation.

PendingSwaps

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Traits

SwapAction
Definition of a pending atomic swap action. It contains the following three phrases:

Trait
Atomic swap’s pallet configuration trait.

Type Definitions

Event
RawEvent specialized for the configuration Trait

HashedProof
Hashed proof type.